#8.14 Bugfix
## 문제점

1. 현재에는 user 가 `.save()` 될 때마다 Password가 Hash 되고 있다.
    - 한번 해싱된 hash 를 반복적으로 해싱 하기 때문에, 초기 설정한 비밀번호로 로그인 할 수 없다.
2. Owner가 아닐 경우 수정 / 삭제 버튼을 가려놓기만 한 것이지, 누구든 `route`를 통해서 접근 할 수 있다.

## 해결

1. 특정 상황(비밀번호가 **생성**되거나 **수정** 될 때) 에서만 Password를 Hash 하여야 한다.
    - `bcrypt` middle ware 수정

        ```jsx
        userSchema.per('save', async function(){
        	if(this.isModified('password')){
        		this.password = await bcrypt.hash(this.password, 5);
        	}
        });
        ```

        - `this[.isModified()](https://mongoosejs.com/docs/api/document.html#document_Document-isModified)` - property 중 하나라도 수정 되면 `true` 반환. else `false` 반환.
2. route 자체를 보호 하여야 한다.
    - Video Controller 수정

        ```jsx
        if(String(video.owner) !== String(req.session.user._id)){
        	return res.status(403).redirect('/');
        }
        ```

        - 현재 로그인 된 User의 ID와 해당 동영상 소유자의 ID를 비교한다.
        - JavaScript 는 Type도 비교 하므로, `String()` 으로 Type변환 해준다.
        - **[403** HTTP Error](https://developer.mozilla.org/ko/docs/Web/HTTP/Status/403) : **Forbidden.** 서버에 요청이 전달되었지만, 권한 때문에 거절되었다는 것을 의미
        - 현재는 owner의 id 값만 필요하므로 User의 전체 Object가 필요하지 않다.

            → `.populate()`는 필요 없다. 필요 없는 DB 요청으로 DB를 혹사 시키면 안된다.

+++ src/controllers/videoControllers.js
@@ -45,19 +45,35 @@ export const postUpload = async (req, res) => {
     }
 };
 export const getEdit = async (req, res) => {
-    const {id} = req.params;
+    const {
+        params: { id },
+        session: {
+            user: { _id }
+        }
+    } = req;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render('common/404', {pageTitle: 'Video not Found'});
     }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect('/');
+    }
     return res.render('videos/edit', {pageTitle: `Editing : ${video.title}`, video});
 };
 export const postEdit = async (req, res) => {
-    const {id} = req.params;
-    const { title, description, hashtags } = req.body;
+    const {
+        params: { id },
+        body: {title, description, hashtags},
+        session: {
+            user: { _id },
+        },
+    } = req;
     const video = await Video.exists({_id : id});
     if(!video){
         return res.render('common/404', {pageTitle: 'Video not Found'});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect('/');
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -67,7 +83,19 @@ export const postEdit = async (req, res) => {
     return res.redirect(`/videos/${id}`)
 };
 export const deleteVideo = async (req, res) => {
-    const {id} = req.params;
+    const {
+        params: { id },
+        session: {
+            user: { _id },
+        },
+    } = req;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render('common/404', {pageTitle: 'Video not Found'});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect('/');
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect('/');
 };
